refactor: inline two-factor authentication composable logic into TwoFactor component to simplify structure
